Move all client messages into `firezone-tunnel`
